ターン制の概念導入
