+++ position.rb
@@ -1,0 +1,6 @@
+class Position
+  def initialize(count,turn)
+    @count = count
+    @turn = turn
+  end
+end
